+++ examples/udp_sender_receiver/udp_receiver.c
@@ -172,7 +172,7 @@ int main(int argc, char* argv[])
         kodo_read_payload(decoder, payload);
 
         if (kodo_has_partial_decoding_tracker(decoder) &&
-            kodo_is_partial_complete(decoder))
+            kodo_is_partially_complete(decoder))
         {
             uint32_t i = 0;
             for (; i < kodo_symbols(decoder); ++i)